Logging buffer.
