+++ packages/konsole-app/src/index.ts
@@ -6,6 +6,7 @@ export * from './components';
 export * from './containers';
 export * from './hooks';
 export * from './icons';
+export * from './logging';
 export * from './panels';
 export * from './testing';
 export * from './theme';
